[osDependency] Centos support
  - 기능
    - hypermachine provisioning시 centos 지원
  - 테스트
    - master/worker 생성 및 삭제 테스트
