+++ scripts/centos/docker.sh
@@ -1,0 +1,3 @@
+yum -y install docker docker-registry
+systemctl enable docker.service
+systemctl start docker.service
