+++ utils/const.go
@@ -8,5 +8,6 @@ const (
 	LabelClusterRoleMaster     = "cluster.x-k8s.io/control-plane"
 	LabelClusterRoleWorker     = "cluster.x-k8s.io/deployment-name"
 
-	SetProviderID2KubeletConfig = "sed -i 's/--config=\\/var\\/lib\\/kubelet\\/config.yaml/--config=\\/var\\/lib\\/kubelet\\/config.yaml --provider-id hyper:\\/\\/\\/bmpName/' /etc/systemd/system/kubelet.service.d/10-kubeadm.conf"
+	SetUbuntuProviderID2KubeletConfig = "sed -i 's/--config=\\/var\\/lib\\/kubelet\\/config.yaml/--config=\\/var\\/lib\\/kubelet\\/config.yaml --provider-id hyper:\\/\\/\\/bmpName/' /etc/systemd/system/kubelet.service.d/10-kubeadm.conf"
+	SetCentosProviderID2KubeletConfig = "sed -i 's/--config=\\/var\\/lib\\/kubelet\\/config.yaml/--config=\\/var\\/lib\\/kubelet\\/config.yaml --provider-id hyper:\\/\\/\\/bmpName/' /usr/lib/systemd/system/kubelet.service.d/10-kubeadm.conf"
 )
